code cleanups
* removed not used imports
* removed not needed casts
* removed doublicated ";"

+++ driver-core/src/main/java/com/datastax/driver/core/utils/UUIDs.java
@@ -86,7 +86,7 @@ public final class UUIDs {
 
             long node = 0;
             for (int i = 0; i < 6; i++)
-                node |= (0x00000000000000ffL & (long)hash[i]) << (i*8);
+                node |= (0x00000000000000ffL & hash[i]) << (i*8);
             // Since we don't use the mac address, the spec says that multicast
             // bit (least significant bit of the first octet of the node ID) must be 1.
             return node | 0x0000010000000000L;
